ci: remove pnpm filters and use working dirs
Each job either acts on the "global" working directory or
a sub-directory where the different NodeJS projects are inside (there
is only one exception for the docs generation). Therefore it barely
makes sense to use pnpm filters or changing directories on steps level.
Rather the whole jobs uses the sub-directory as starting point.
This also allows for some simplifications of the storage of artifacts.