ci: make use of workflow 'when' property
We have multiple workflows in the CircleCI pipline that should run on
different conditions. So far the approach is to apply a filter property
on the install job of each workflow. Since all other jobs always depend
at least on this job, this makes sure no other job is running if not
this one. Furthermore this does the assumption that always the full
workflow should get executed (although further filters could be added of
course).
While this is a working solution, the code does not really express the
intention as described in the introduction. But what can is the `when`
property of workflows. They do exactly define a condition when to run
the according workflow or not. As downside this does not support regular
expressions. But with the current setup this is not an issue. The only
pattern we use is for release tags. But since tags are handled special
in general, this does not break the usage of the when condition. Finally
the nightly workflow still uses its trigger.
Besides a better logical description, we escape the boundaries of the
filters. This means we can write higher level boolean expression that
have access to the whole pipeline. So for example we could decide to
introduce a new variable that allows to trigger a workflow, although the
other conditions like git branch etc are not met. This comes handy when
testing changes to the pipeline with manual runs via the API, avoiding
the need to overwrite triggers temporally or force specific git events.